Add typescript to server
